updated extraction tutorial
